メーカーマスタに優先する発注先を登録している発注先を発注先マスタで削除しようとした場合の警告メッセージを出力する処理の追加

+++ budget/app/Http/Controllers/SupplierController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Supplier;
+use App\Maker;
 use App\Rules\Exists;
 use App\Exceptions\ExclusiveLockException;
 
@@ -143,6 +144,11 @@ class SupplierController extends Controller
      */
     public function destroy($id)
     {
+        //メーカーマスタの優先する発注先に登録されていれば、メッセージを出力
+        $exists = Maker::where('MainSupplierId',$id)->exists();
+        if ($exists) {
+            return redirect()->back()->with('MainSupplierIdError', '優先する発注先に指定しているメーカーがある為、削除できません');
+        }
         $Supplier = Supplier::lockForUpdate()->withTrashed()->find($id);
         $Supplier->delete();
